Reorganize directory structure for override dir
This allows the repo to simply be cloned into the override directory of
the ToEE install. Previously I was trying to clone it elsewhere, and use
symlinks into the override directory to manage actually using parts of
it, but it seems like Temple+ has trouble actually loading content from
symlinks for some of the extender files and such. So it's preferable to
have the actual files/repo reside in the override directory.